hw7/attempt1/requirements.txt
Flask
requests_oauthlib
requests
python-dotenv
